+++ lib/rust/parser/debug/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "enso-parser-debug"
+version = "0.1.0"
+authors = ["Enso Team <enso-dev@enso.org>"]
+edition = "2021"
+description = "Enso parser debugging tools."
+readme = "README.md"
+homepage = "https://github.com/enso-org/enso"
+repository = "https://github.com/enso-org/enso"
+license-file = "../../LICENSE"
+
+[dependencies]
+enso-parser = { path = "../" }
+enso-metamodel = { path = "../../metamodel", features = ["rust"] }
+enso-metamodel-lexpr = { path = "../../metamodel/lexpr" }
+enso-reflect = { path = "../../reflect" }
+lexpr = "0.2.6"
+serde = { version = "1.0", features = ["derive"] }
